updated rendering, currently ugly
