+++ server/routes/listRouter.js
@@ -15,10 +15,7 @@ router.get('/getAllProd', listController.getAll, (req, res) => {
 // post shopping list- click submit (update or post)
 router.post('/submitList', listController.submitList, mazeController.findPath, (req, res) => {
   if (res.locals.shoppingList && res.locals.layoutWithProductsAndPath) {
-    return res.status(200).json({
-      // shoppingList: res.locals.shoppingList, 
-      updatedLayout: res.locals.layoutWithProductsAndPath
-    });
+    return res.status(200).json(res.locals.layoutWithProductsAndPath);
   } else {
     return res.status(500);
   }
